+++ emsa2.h
@@ -61,13 +61,13 @@ class CRYPTOPP_DLL EMSA2Pad : public EMSA2HashIdLookup<PK_DeterministicSignature
 public:
 	static const char * CRYPTOPP_API StaticAlgorithmName() {return "EMSA2";}
 	
-	unsigned int MinRepresentativeBitLength(unsigned int hashIdentifierLength, unsigned int digestLength) const
+	size_t MinRepresentativeBitLength(size_t hashIdentifierLength, size_t digestLength) const
 		{return 8*digestLength + 31;}
 
 	void ComputeMessageRepresentative(RandomNumberGenerator &rng, 
-		const byte *recoverableMessage, unsigned int recoverableMessageLength,
+		const byte *recoverableMessage, size_t recoverableMessageLength,
 		HashTransformation &hash, HashIdentifier hashIdentifier, bool messageEmpty,
-		byte *representative, unsigned int representativeBitLength) const;
+		byte *representative, size_t representativeBitLength) const;
 };
 
 //! EMSA2, for use with RWSS and RSA_ISO
